- Rest Class: new method isAjax() which returns true if a jQuery Ajax request was made. To be used in conjunction with getRequestedMethod so you'll know if the the requested method was made with ajax or not.
- Growl Class: new helper class to manage session based messaging to the frontend. You can send a message to the session object, get that message(clearing the session) and check if there is a message.
The idea for this class is that you can put a message on the session object, do a redirect and assign the session message to the view exactly once, so that if the user reloads it will only see the message once.
Of course, all the processing of the message will still have to me made on the frontend and controller.