Dispatch param bug corrected
Dispatch was not passing an associative array of key=>value params and
was instead an array of function arguments. By wrapping the route array
inside an array we force only one param passed, an array of key=>value
params which can then be tested. We loose some flexibility in defining
param defaults but that was bad practice anyway as we should explicitly
validate each and every param.