Changes to Template Class
+ Remove logic that used the TEMPLATES and TEMPLATES_URI constants all
over the place. Now you only need them for default values if you don't
define the template_dir and template_baseuri.
+ You can call any LessPHP methods from the template class by including
an array with the methods to call as keys and the params as values. ex:
$this->tpl->less_options(array('setImportDir',array('less/')). Initially
I thought about limiting what kind of commands we should allow, but that
seemed too restrictive.